+++ src/tiled/scriptmanager.h
@@ -21,6 +21,7 @@
 #pragma once
 
 #include "filesystemwatcher.h"
+#include "logginginterface.h"
 #include "tilededitor_global.h"
 
 #include <QJSValue>
@@ -58,6 +59,27 @@ public:
         {}
     };
 
+    /**
+     * A line of output emitted while a script was being evaluated.
+     */
+    struct OutputLine {
+        LoggingInterface::OutputType type;
+        QString text;
+    };
+
+    /**
+     * The result of evaluating a script with evaluateCaptured().
+     */
+    struct EvaluationResult {
+        QString tempName;           // global name assigned to the result ($0, $1, ...)
+        QString result;             // string representation of the result
+        QString error;              // error message, if evaluation failed
+        QList<OutputLine> output;   // output emitted during evaluation
+        bool hasResult = false;     // false when the result was undefined or an error
+
+        bool hasError() const { return !error.isNull(); }
+    };
+
     static ScriptManager &instance();
     static void deleteInstance();
 
@@ -73,6 +95,9 @@ public:
     QJSValue evaluate(const QString &program,
                       const QString &fileName = QString(), int lineNumber = 1);
 
+    EvaluationResult evaluateCaptured(const QString &program,
+                                      const QString &fileName = QString());
+
     void evaluateFileOrLoadModule(const QString &fileName);
 
     /**
@@ -82,6 +107,7 @@ public:
     QString createTempValue(const QJSValue &value);
 
     bool checkError(QJSValue value, const QString &program = QString());
+    static QString errorString(const QJSValue &value, const QString &program = QString());
     void throwError(const QString &message);
     void throwNullArgError(int argNumber);